requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0  # Добавьте эту строку
python-multipart==0.0.6
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.11.1